backend/ChatBot/requirements.txt
groq>=0.9.0
flask>=3.0.0
pymongo>=4.6.0
python-dotenv>=1.0.1
